+++ src/data/cells/cellRenderer.ts
@@ -244,7 +244,7 @@ export const ReplaceCellMacros = (
   //
   // Replace $__cell with this cell's content $__cell word boundary
   //
-  formatted = formatted.replace(/\$__cell\b/g, cellContent);
+  formatted = formatted.replace(/\$__cell\b/g, () => cellContent);
 
   //
   // process $__cell_N